pyproject.toml
[project]
name = "rag-app-qa-with-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.1",
    "pymupdf>=1.27.1",
    "pyowm>=3.5.0",
    "pypdf>=6.7.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.4.0",
    "streamlit>=1.54.0",
    "wikipedia>=1.4.0",
]
